final code

+++ CompileCode.js
@@ -1,7 +1,7 @@
 const { exec } = require('child_process');
 const fs = require('fs');
 
-const CompileCode = (code, language) => {
+const CompileCode = (code, language, input = '') => {  // Default value for input
   return new Promise((resolve, reject) => {
     const filename = `temp-code.${language}`;
     const outputFilename = "temp-output";
@@ -13,25 +13,29 @@ const CompileCode = (code, language) => {
       compileCommand = `g++ -std=c++17 ${filename} -o ${outputFilename}`;  
     } else if (language === 'c') {
       compileCommand = `gcc ${filename} -o ${outputFilename}`;
-    } // Add more language conditions here
+    }
 
     exec(compileCommand, (error, stdout, stderr) => {
       if (error || stderr) {
-        fs.unlinkSync(filename);
-        reject(error || stderr);
-      } else {
-        exec(`./${outputFilename}`, (error, stdout, stderr) => {
-          fs.unlinkSync(filename);
-          fs.unlinkSync(outputFilename);
-          if (error || stderr) {
-            reject(error || stderr);
-          } else {
-            resolve(stdout);
-          }
-        });
+        if (fs.existsSync(filename)) fs.unlinkSync(filename);
+        if (fs.existsSync(outputFilename)) fs.unlinkSync(outputFilename);  // Check if file exists before deleting
+        reject({ type: 'compile', message: error || stderr });
+        return;
       }
+
+      const execCommand = input ? `echo "${input}" | ./${outputFilename}` : `./${outputFilename}`;  // Conditionally add echo
+
+      exec(execCommand, (error, stdout, stderr) => {
+        if (fs.existsSync(filename)) fs.unlinkSync(filename);
+        if (fs.existsSync(outputFilename)) fs.unlinkSync(outputFilename);  // Check if file exists before deleting
+        if (error || stderr) {
+          reject({ type: 'runtime', message: error || stderr });
+        } else {
+          resolve(stdout);
+        }
+      });
     });
   });
 };
 
-module.exports = CompileCode;
+module.exports=CompileCode
